+++ bin/plot_peak_widths.R
@@ -2,6 +2,7 @@
 library(dplyr)
 library(ggplot2)
 library(readr)
+library(scales)
 
 set.seed(20230926)
 args <- commandArgs(trailingOnly = TRUE)
@@ -10,30 +11,57 @@ if (length(args) < 1) {
 }
 tsv_filename <- args[1]
 peak_dat <- read_tsv(tsv_filename) %>%
-  mutate(peak_width = chromEnd - chromStart)
-xmin <- peak_dat %>%
-  pull(peak_width) %>%
-  min()
-xmax <- peak_dat %>%
-  pull(peak_width) %>%
-  max()
+  mutate(
+    peak_width = chromEnd - chromStart,
+    peak_type = case_when(
+      tool == "sicer" | tool == "macs_broad" ~ "broad",
+      tool == "gem" | tool == "macs_narrow" ~ "narrow",
+      TRUE ~ NA_character_
+    )
+  )
 
-peak_widths_hist <- peak_dat %>%
-  ggplot(aes(peak_width, fill = tool)) +
-  geom_histogram(alpha = 0.7, position = "identity") +
-  scale_x_log10(
-    limits = c(xmin - 10^2, xmax + 10^2),
-    labels = scales::label_log(digits = 2)
-  ) +
-  scale_fill_viridis_d() +
-  guides(fill = guide_legend(
-    label.position = "bottom",
-    title = "Peak caller",
-    title.position = "top"
-  )) +
-  facet_wrap("sample_id") +
-  labs(x = expression("" * log[10] * " Peak Widths")) +
-  theme_bw() +
-  theme(legend.position = "top")
+tool_colors <- viridisLite::viridis(n = 4, alpha = 0.7, begin = 0, end = 1, direction = 1, option = "D")
+names(tool_colors) <- peak_dat %>%
+  pull(tool) %>%
+  unique()
 
-ggsave(filename = "peak_widths_histogram.png", plot = peak_widths_hist, device = "png")
+plot_hist <- function(peak_dat) {
+  xmin <- peak_dat %>%
+    pull(peak_width) %>%
+    min()
+  xmax <- peak_dat %>%
+    pull(peak_width) %>%
+    max()
+  peak_dat %>%
+    ggplot(aes(peak_width, fill = tool)) +
+    geom_histogram(alpha = 0.7, position = "identity") +
+    scale_fill_manual(
+      values = tool_colors,
+      breaks = names(tool_colors)
+    ) +
+    guides(fill = guide_legend(
+      label.position = "bottom",
+      title = "Peak caller",
+      title.position = "top"
+    )) +
+    facet_wrap(~sample_id) +
+    labs(x = expression("" * log[10] * " Peak Widths")) +
+    theme_bw() +
+    theme(legend.position = "top")
+}
+
+hist_broad <- peak_dat %>%
+  filter(peak_type == "broad") %>%
+  plot_hist()
+hist_narrow <- peak_dat %>%
+  filter(peak_type == "narrow") %>%
+  plot_hist()
+
+ggsave(
+  filename = "peak_widths_broad_histogram.png", plot = hist_broad,
+  device = "png", dpi = 300, height = 4, width = 6
+)
+ggsave(
+  filename = "peak_widths_narrow_histogram.png", plot = hist_narrow,
+  device = "png", dpi = 300, height = 4, width = 6
+)